Post Featured Image: Optimize store subscriptions
